Implement disabling cave and ravine generation, clean up some code ramaining after generation pipeline
