--- src/main/java/cubicchunks/worldgen/generator/custom/CustomPopulator.java
@@ -1,72 +1,0 @@
-/*
- *  This file is part of Cubic Chunks Mod, licensed under the MIT License (MIT).
- *
- *  Copyright (c) 2015 contributors
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in
- *  all copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *  THE SOFTWARE.
- */
-package cubicchunks.worldgen.generator.custom;
-
-import cubicchunks.api.worldgen.biome.CubicBiome;
-import cubicchunks.api.worldgen.populator.CubePopulatorEvent;
-import cubicchunks.api.worldgen.populator.ICubicPopulator;
-import cubicchunks.util.Coords;
-import cubicchunks.util.CubePos;
-import cubicchunks.world.ICubicWorld;
-import cubicchunks.world.cube.Cube;
-import cubicchunks.worldgen.generator.CubeGeneratorsRegistry;
-import mcp.MethodsReturnNonnullByDefault;
-import net.minecraftforge.common.MinecraftForge;
-
-import java.util.Random;
-
-import javax.annotation.ParametersAreNonnullByDefault;
-
-@ParametersAreNonnullByDefault
-@MethodsReturnNonnullByDefault
-public class CustomPopulator {
-
-    private ICubicWorld world;
-
-    public CustomPopulator(ICubicWorld world) {
-        this.world = world;
-    }
-
-    public void populate(Cube cube) {
-        /**
-         * If event is not canceled we will use default biome decorators and
-         * cube populators from registry.
-         **/
-        if (!MinecraftForge.EVENT_BUS.post(new CubePopulatorEvent(world, cube))) {
-            CubicBiome biome = CubicBiome.getCubic(cube.getCubicWorld().getBiome(Coords.getCubeCenter(cube)));
-
-            CubePos pos = cube.getCoords();
-            // For surface generators we should actually use special RNG with
-            // seed
-            // that depends only in world seed and cube X/Z
-            // but using this for surface generation doesn't cause any
-            // noticeable issues
-            Random rand = new Random(cube.cubeRandomSeed());
-
-            ICubicPopulator decorator = biome.getDecorator();
-            decorator.generate(world, rand, pos, biome);
-            CubeGeneratorsRegistry.generateWorld(world, rand, pos, biome);
-        }
-    }
-}
